==> hourse_data/1.5.sum.ipynb <==
## データを追加したい時にdfとdfを結合

import pandas as pd
import time

pd.set_option('display.max_rows',5)
pd.options.display.width = 200
pd.options.display.column_space = 100

df1 = pd.read_csv("1.2.test1.csv")
df2 = pd.read_csv("1.2.test2.csv")
# df2 = pd.read_pickle("df.pickle")
print(len(df1))
print(len(df2))

df1 = df1.iloc[:,:15]
df2 = df2.iloc[:,:15]

# pd.set_option('display.max_rows',None)

df = pd.concat(objs=[df1, df2])
# df

len(df)

df.to_pickle("1.7.df.pickle")
time.sleep(2)

pd.set_option('display.max_rows',None)

df = pd.read_pickle('1.7.df.pickle')

pd.options.display.max_rows = None
for i in df.columns:
  print(df[i].value_counts())
