==> hourse_data/2.0.add.詳細.ipynb <==
## 3.日付データやその他情報を追加

import pandas as pd
import requests
from bs4 import BeautifulSoup
import pandas as pd
import re
import time
from tqdm import tqdm

 データの読み込み

pd.set_option('display.max_rows',5)
pd.options.display.width = 500
pd.options.display.column_space = 200

df = pd.read_pickle('1.7.df.pickle')
race_id_list = df.index.unique()
str_id_list = list(map(str,race_id_list))
str_id_list

race_infos = {}
def add_data(race_id_list):
  for race_id in tqdm(race_id_list):
    url = "https://race.netkeiba.com/race/result.html?race_id=" + race_id
    html = requests.get(url)
    html.encoding= "EUC_JP"
    soup = BeautifulSoup(html.text, "html.parser")
    time.sleep(1)
  
    日付 = soup.find("dd",class_='Active')
    日付 = 日付.find("a").get_text()
    レース情報 = soup.find("div", class_="RaceList_Item02") 
    # 距離 = レース情報.find("div",class_="RaceData01").find("span").get_text()
    # 馬場 = レース情報.find("div",class_="RaceData01").find("span",class_="Item03").get_text()
    レース環境 = レース情報.find("div",class_="RaceData01").get_text()
    info = re.findall(r"\w+",レース環境)
    info[0] = info[2][:1]
    info[1] = info[2][1:]
    info.pop(2)
    info_dict = {}
    # print(info)
    info_dict["date"] = f"{race_id[:4]}年{日付[:-3]}"

    for text in info:
      # print(text)
      if text in ["芝","ダ"]:
        # print("芝付き")
        info_dict["race_type"] = text
      if "障" in text:
        info_dict["race_type"] = "障害"
      if "m" in text:
        # print("mつき")
        info_dict["course_ren"] = re.findall(r"\d+",text)[0]
      if text in ["雨","小雨","小雪","雪","晴","曇"]:
        # print("天気")
        info_dict["weather"] = text
      # if "年" in text:
      #   info_dict["data"] = text
      if text in ["良","不良","重","稍","不"]:
        # print("馬場")
        info_dict["ground_state"] = text

    race_infos[race_id] = info_dict
  return race_infos

results = add_data(str_id_list)

----

## 4.dataの形を成形して元のデータと結合

race_infos_df = pd.DataFrame(results).T
# race_infos_df

結合するdataframeの形が一致しないと結合できないので注意

df.index = df.index.astype(str)

race_infos_df.index = race_infos_df.index.astype(str)

strに変わったか確認

type(df.index[0])

pd.set_option('display.max_rows',None)

inner_df = df.merge(race_infos_df,left_index=True,right_index=True,how="inner")

----

# df = pd.read_pickle("2.2.add.日付_df.pickle")
inner_df.to_csv('2.2.add.日付_df.csv')

pd.options.display.max_rows = None
for i in df.columns:
  print(df[i].value_counts())

df[df["人気"] == "1:02.7"].index
