==> hourse_data/4.clean.data.ipynb <==
import pandas as pd
import time

pd.options.display.max_rows = None
pd.options.display.max_columns = 100
pd.options.display.max_colwidth = 20

horse_vs_df = pd.read_csv('3.4.horse_vs_data.csv')
horse_vs_df = horse_vs_df.iloc[:,:28]
df = pd.read_csv('3.1.df_add_id.csv')
print(len(horse_vs_df))
print(len(df))
# time.sleep(10)

着順の１位　＝　NAN 走ってないうま　＝　NAN これをくべつする

## メインデータの前処理

def preprocessing(df):
  # 着順
  df["着順"] = df["着順"].map(lambda x:str(x).replace('除外', "21"))
  df["着順"] = df["着順"].map(lambda x:str(x).replace('取消', "21"))
  df["着順"] = df["着順"].map(lambda x:str(x).replace('中止', "21"))
  df["着順"] = df["着順"].astype(int)
  # 性齢
  df["性"] = df["性齢"].map(lambda x:str(x)[0])
  df["齢"] = df["性齢"].map(lambda x:str(x)[1:]).astype(int)
  # 人気
  df["人気"] = df["人気"].map(lambda x:str(x).replace('nan', "21"))
  df["人気"] = df["人気"].map(lambda x:str(x).replace('.0', ''))
  df["人気"] = df["人気"].astype(int)
  # タイム
  df['タイム'] = df['タイム'].fillna("9:99.9")
  # 着差
  # df.loc[df['着差'] == "-", '着差'] = "除外"
  # df["着差"] = df["着差"].fillna("１着")
  # オッズ
  df["単勝オッズ"] = df["単勝オッズ"].fillna(0)
  df["単勝オッズ"] = df["単勝オッズ"].astype(float)
  # 後3F
  df["後3F"] = df["後3F"].fillna(99.9)
  df["後3F"] = df["後3F"].astype(float)
  # コーナー通過順
  df["コーナー通過順"] = df["コーナー通過順"].fillna(0-0-0-0)
  # 馬体重
  df["馬体重(増減)"] = df["馬体重(増減)"].fillna("999(999)")
  df["体重"] = df["馬体重(増減)"].str.split('(',expand=True)[0].astype(int)
  df["体重変化"] = df["馬体重(増減)"].str.split('(',expand=True)[1].str[:-1]
  df.drop(["馬体重(増減)",'性齢'],axis=1, inplace=True)
  # 日付
  df["date"] = pd.to_datetime(df["date"],format="%Y年%m月%d日")

  return df  

## HorseResultsクラス

class HorseResults:
  def __init__(self,horse_vs_df):
    self.horse_results = horse_vs_df[["日付","着順","賞金"]]
    
  def preprocessing(self):
    df = self.horse_results.copy()
    # 着順
    df["着順"] = df["着順"].map(lambda x:str(x).replace('.0', ''))
    df["着順"] = pd.to_numeric(df["着順"],errors="coerce")
    df.dropna(subset=["着順"],inplace=True)
    df["着順"] = df["着順"].astype(int)
    # 日付
    df["date"] = pd.to_datetime(df["日付"])
    df.drop(["日付"],axis=1, inplace=True)
    
    self.horse_results = df


hr = HorseResults(horse_vs_df)
hr.preprocessing()
hr.horse_results.head()
# hr.horse_results.index = hr.horse_results.index.map(lambda x:str(x))

１度目の実行で必ずエラーが出る (なぜか１度目の実行でnanが表示されないから)

results_p = preprocessing(df)
len(results_p)

# pd.options.display.max_rows = None
# for i in df.columns:
#   print(df[i].value_counts())
# results_p.info()

results_p.head()

sample_data = results_p["date"].iloc[1]
sample_data

sample_dataの日付に走った馬のリスト


pd.options.display.max_rows = None

horse_id_list = results_p[results_p["date"] == sample_data]["horse_id"]
horse_id_list

"3.4.horse_vs_data.csv"に取得できていないデータがたまにあるので、エラーが出たら再取得してコピペすれば解決する "3.0.add.id.ipynb"

target_df = hr.horse_results.loc[horse_id_list]
target_df
target_df[target_df["date"] < sample_data].groupby(level=0)[["着順","賞金"]].mean()
